@@ -1,6 +1,6 @@
 const EMAIL_REGEX = /^[a-z0-9!#$%&'*+/=?^_`{|}~.-]+@[a-z0-9-]+(\.[a-z0-9-]+)*$/i;
 
-export function email( val )
+export function email( val: string )
 {
 	return EMAIL_REGEX.test( val );
 };
